requirements.txt
socksio
streamlit
openai==v0.28.1
python-dotenv
